proto/requirements.txt
# Python prototype dependencies
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
sentencepiece>=0.2.0
llama-cpp-python>=0.3.0
huggingface-hub>=0.20.0
ruff>=0.9.0
ty>=0.0.12,<0.1.0
